@@ -4,16 +4,12 @@ require "sidekiq/web"
 require "sidekiq/cron/web"
 
 Rails.application.routes.draw do
-  authenticate :user, ->(u) { u.admin? } do
-    mount Sidekiq::Web => "/sidekiq"
-  end
-
   mount LetterOpenerWeb::Engine, at: "/letter_opener" if Rails.env.development?
 
-  namespace :admin do
-    resources :iframe, only: [:index]
+  authenticate :user, ->(u) { u.admin? } do
+    mount Sidekiq::Web => "/sidekiq"
+    mount ActiveHashcash::Engine, at: "hashcash"
   end
 
   mount Decidim::Core::Engine => "/"
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
