@@ -10,19 +10,35 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2024_09_23_110435) do
-
+ActiveRecord::Schema[7.0].define(version: 2025_05_16_090459) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "ltree"
   enable_extension "pg_trgm"
   enable_extension "plpgsql"
+
+  create_table "active_hashcash_stamps", force: :cascade do |t|
+    t.string "version", null: false
+    t.integer "bits", null: false
+    t.date "date", null: false
+    t.string "resource", null: false
+    t.string "ext", null: false
+    t.string "rand", null: false
+    t.string "counter", null: false
+    t.string "request_path"
+    t.string "ip_address"
+    t.jsonb "context"
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
+    t.index ["counter", "rand", "date", "resource", "bits", "version", "ext"], name: "index_active_hashcash_stamps_unique", unique: true
+    t.index ["ip_address", "created_at"], name: "index_active_hashcash_stamps_on_ip_address_and_created_at", where: "(ip_address IS NOT NULL)"
+  end
 
   create_table "active_storage_attachments", force: :cascade do |t|
     t.string "name", null: false
     t.string "record_type", null: false
     t.bigint "record_id", null: false
     t.bigint "blob_id", null: false
-    t.datetime "created_at", null: false
+    t.datetime "created_at", precision: nil, null: false
     t.index ["blob_id"], name: "index_active_storage_attachments_on_blob_id"
     t.index ["record_type", "record_id", "name", "blob_id"], name: "index_active_storage_attachments_uniqueness", unique: true
   end
@@ -34,7 +50,7 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.text "metadata"
     t.bigint "byte_size", null: false
     t.string "checksum", null: false
-    t.datetime "created_at", null: false
+    t.datetime "created_at", precision: nil, null: false
     t.string "service_name", null: false
     t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
   end
@@ -56,8 +72,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "decidim_accountability_status_id"
     t.integer "decidim_component_id"
     t.integer "decidim_scope_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "children_count", default: 0
     t.float "weight", default: 1.0
     t.string "external_id"
@@ -73,8 +89,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "key"
     t.jsonb "name"
     t.integer "decidim_component_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "description"
     t.integer "progress"
     t.index ["decidim_component_id"], name: "index_decidim_accountability_statuses_on_decidim_component_id"
@@ -84,8 +100,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.date "entry_date"
     t.jsonb "description"
     t.integer "decidim_accountability_result_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "title"
     t.index ["decidim_accountability_result_id"], name: "index_decidim_accountability_timeline_entries_on_results_id"
     t.index ["entry_date"], name: "index_decidim_accountability_timeline_entries_on_entry_date"
@@ -101,8 +117,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "participatory_space_id"
     t.string "action", null: false
     t.jsonb "extra"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "version_id"
     t.string "visibility", default: "admin-only"
     t.integer "decidim_scope_id"
@@ -125,49 +141,13 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "decidim_amendable_id"
     t.string "decidim_emendation_type"
     t.bigint "decidim_emendation_id"
-    t.string "state"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
+    t.integer "state", default: 0, null: false
     t.index ["decidim_amendable_id", "decidim_amendable_type"], name: "index_on_amendable"
     t.index ["decidim_emendation_id"], name: "index_decidim_amendments_on_decidim_emendation_id"
     t.index ["decidim_user_id", "decidim_amendable_id", "decidim_amendable_type"], name: "index_on_amender_and_amendable"
     t.index ["decidim_user_id"], name: "index_decidim_amendments_on_decidim_user_id"
-    t.index ["state"], name: "index_decidim_amendments_on_state"
-  end
-
-  create_table "decidim_anonymous_codes_groups", force: :cascade do |t|
-    t.jsonb "title"
-    t.datetime "expires_at"
-    t.boolean "active", default: true, null: false
-    t.integer "max_reuses", default: 1, null: false
-    t.integer "tokens_count", default: 0, null: false
-    t.string "resource_type"
-    t.bigint "resource_id"
-    t.bigint "decidim_organization_id", null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
-    t.index ["decidim_organization_id"], name: "decidim_anonymous_codes_groups_on_organization"
-    t.index ["resource_type", "resource_id"], name: "decidim_anonymous_codes_groups_on_resource"
-  end
-
-  create_table "decidim_anonymous_codes_token_resources", force: :cascade do |t|
-    t.bigint "token_id", null: false
-    t.string "resource_type", null: false
-    t.bigint "resource_id", null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
-    t.index ["resource_type", "resource_id"], name: "decidim_anonymous_codes_token_resources_on_resource"
-    t.index ["token_id"], name: "decidim_anonymous_codes_token_resources_on_token"
-  end
-
-  create_table "decidim_anonymous_codes_tokens", force: :cascade do |t|
-    t.string "token", null: false
-    t.integer "usage_count", default: 0, null: false
-    t.bigint "group_id", null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
-    t.index ["group_id"], name: "decidim_anonymous_codes_tokens_on_group"
-    t.index ["token", "group_id"], name: "index_anonymous_codes_token_group_uniqueness", unique: true
   end
 
   create_table "decidim_area_types", force: :cascade do |t|
@@ -181,8 +161,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "name"
     t.bigint "area_type_id"
     t.bigint "decidim_organization_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["area_type_id"], name: "index_decidim_areas_on_area_type_id"
     t.index ["decidim_organization_id"], name: "index_decidim_areas_on_decidim_organization_id"
   end
@@ -191,8 +171,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "slug", null: false
     t.string "hashtag"
     t.integer "decidim_organization_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "title", null: false
     t.jsonb "subtitle", null: false
     t.jsonb "short_description", null: false
@@ -200,14 +180,14 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "hero_image"
     t.string "banner_image"
     t.boolean "promoted", default: false
-    t.datetime "published_at"
+    t.datetime "published_at", precision: nil
     t.jsonb "developer_group"
     t.jsonb "meta_scope"
     t.jsonb "local_area"
     t.jsonb "target"
     t.jsonb "participatory_scope"
     t.jsonb "participatory_structure"
-    t.boolean "show_statistics", default: false
+    t.boolean "show_statistics", default: true
     t.integer "decidim_scope_id"
     t.boolean "scopes_enabled", default: true, null: false
     t.boolean "private_space", default: false
@@ -254,8 +234,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_assemblies_types", force: :cascade do |t|
     t.jsonb "title", null: false
     t.integer "decidim_organization_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_organization_id"], name: "index_decidim_assemblies_types_on_decidim_organization_id"
   end
 
@@ -270,8 +250,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "position"
     t.string "position_other"
     t.date "ceased_date"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.bigint "decidim_user_id"
     t.index ["decidim_assembly_id"], name: "index_decidim_assembly_members_on_decidim_assembly_id"
     t.index ["decidim_user_id"], name: "index_decidim_assembly_members_on_decidim_user_id"
@@ -282,8 +262,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "decidim_user_id"
     t.integer "decidim_assembly_id"
     t.string "role"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_assembly_id", "decidim_user_id", "role"], name: "index_unique_user_and_assembly_role", unique: true
     t.index ["decidim_user_id"], name: "index_decidim_assembly_user_roles_on_decidim_user_id"
   end
@@ -304,23 +284,43 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "content_type", null: false
     t.string "file_size", null: false
     t.integer "attached_to_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "attached_to_type", null: false
     t.integer "weight", default: 0, null: false
     t.integer "attachment_collection_id"
+    t.string "link"
     t.index ["attached_to_id", "attached_to_type"], name: "index_decidim_attachments_on_attached_to"
     t.index ["attachment_collection_id"], name: "index_decidim_attachments_on_attachment_collection_id"
+  end
+
+  create_table "decidim_authorization_transfer_records", force: :cascade do |t|
+    t.bigint "transfer_id", null: false
+    t.string "resource_type", null: false
+    t.bigint "resource_id", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.index ["resource_type", "resource_id"], name: "index_decidim_authorization_transfer_records_on_resource"
+    t.index ["transfer_id"], name: "index_decidim_authorization_transfer_records_on_transfer_id"
+  end
+
+  create_table "decidim_authorization_transfers", force: :cascade do |t|
+    t.bigint "user_id", null: false
+    t.bigint "source_user_id", null: false
+    t.bigint "authorization_id", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.index ["authorization_id"], name: "index_decidim_authorization_transfers_on_authorization_id"
+    t.index ["source_user_id"], name: "index_decidim_authorization_transfers_on_source_user_id"
+    t.index ["user_id"], name: "index_decidim_authorization_transfers_on_user_id"
   end
 
   create_table "decidim_authorizations", id: :serial, force: :cascade do |t|
     t.string "name", null: false
     t.jsonb "metadata"
     t.integer "decidim_user_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "unique_id"
-    t.datetime "granted_at"
+    t.datetime "granted_at", precision: nil
     t.jsonb "verification_metadata", default: {}
     t.string "verification_attachment"
     t.index ["decidim_user_id", "name"], name: "index_decidim_authorizations_on_decidim_user_id_and_name", unique: true
@@ -332,8 +332,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "var"
     t.jsonb "value"
     t.integer "decidim_organization_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_organization_id"], name: "index_decidim_awesome_on_decidim_organization_id"
     t.index ["var", "decidim_organization_id"], name: "index_decidim_awesome_organization_var", unique: true
   end
@@ -341,8 +341,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_awesome_config_constraints", force: :cascade do |t|
     t.jsonb "settings"
     t.bigint "decidim_awesome_config_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_awesome_config_id"], name: "decidim_awesome_config_constraints_config"
     t.index ["settings", "decidim_awesome_config_id"], name: "index_decidim_awesome_settings_awesome_config", unique: true
   end
@@ -352,8 +352,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "path"
     t.bigint "decidim_author_id", null: false
     t.bigint "decidim_organization_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_author_id"], name: "decidim_awesome_editor_images_author"
     t.index ["decidim_organization_id"], name: "decidim_awesome_editor_images_constraint_organization"
   end
@@ -362,19 +362,19 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "decidim_proposal_id", null: false
     t.jsonb "vote_weight_totals"
     t.integer "weight_total", default: 0
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.string "private_body"
     t.string "decidim_proposal_type", null: false
-    t.datetime "private_body_updated_at"
+    t.datetime "private_body_updated_at", precision: nil
     t.index ["decidim_proposal_id", "decidim_proposal_type"], name: "index_decidim_awesome_proposal_extra_fields_on_decidim_proposal"
   end
 
   create_table "decidim_awesome_vote_weights", force: :cascade do |t|
     t.bigint "proposal_vote_id", null: false
     t.integer "weight", default: 1, null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["proposal_vote_id"], name: "decidim_awesome_proposals_weights_vote"
   end
 
@@ -382,14 +382,15 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "title"
     t.jsonb "body"
     t.integer "decidim_component_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "decidim_author_id", null: false
     t.string "decidim_author_type", null: false
     t.integer "decidim_user_group_id"
     t.integer "endorsements_count", default: 0, null: false
     t.integer "comments_count", default: 0, null: false
     t.integer "follows_count", default: 0, null: false
+    t.datetime "published_at", precision: nil
     t.index ["decidim_author_id", "decidim_author_type"], name: "index_decidim_blogs_posts_on_decidim_author"
     t.index ["decidim_component_id"], name: "index_decidim_blogs_posts_on_decidim_component_id"
     t.index ["decidim_user_group_id"], name: "index_decidim_blogs_posts_on_decidim_user_group_id"
@@ -401,8 +402,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "description"
     t.integer "total_budget", default: 0
     t.integer "decidim_component_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.bigint "decidim_scope_id"
     t.index ["decidim_component_id"], name: "index_decidim_budgets_budgets_on_decidim_component_id"
     t.index ["decidim_scope_id"], name: "index_decidim_budgets_budgets_on_decidim_scope_id"
@@ -418,9 +419,9 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
 
   create_table "decidim_budgets_orders", id: :serial, force: :cascade do |t|
     t.integer "decidim_user_id"
-    t.datetime "checked_out_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "checked_out_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.bigint "decidim_budgets_budget_id"
     t.index ["decidim_budgets_budget_id"], name: "index_decidim_budgets_orders_on_decidim_budgets_budget_id"
     t.index ["decidim_user_id"], name: "index_decidim_budgets_orders_on_decidim_user_id"
@@ -431,8 +432,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "description"
     t.bigint "budget_amount", null: false
     t.integer "decidim_scope_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "reference"
     t.bigint "decidim_budgets_budget_id"
     t.date "selected_at"
@@ -460,8 +461,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "decidim_category_id", null: false
     t.string "categorizable_type", null: false
     t.bigint "categorizable_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["categorizable_type", "categorizable_id"], name: "decidim_categorizations_categorizable_id_and_type"
     t.index ["decidim_category_id"], name: "index_decidim_categorizations_on_decidim_category_id"
   end
@@ -471,8 +472,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "decidim_user_group_id"
     t.string "coauthorable_type"
     t.bigint "coauthorable_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "decidim_author_type", null: false
     t.index ["coauthorable_type", "coauthorable_id"], name: "index_coauthorable_on_coauthorship"
     t.index ["decidim_author_id", "decidim_author_type"], name: "index_decidim_coauthorships_on_decidim_author"
@@ -483,8 +484,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "weight", null: false
     t.integer "decidim_comment_id", null: false
     t.integer "decidim_author_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "decidim_author_type", null: false
     t.index ["decidim_author_id", "decidim_author_type"], name: "index_decidim_comments_comment_votes_on_decidim_author"
     t.index ["decidim_author_id"], name: "decidim_comments_comment_vote_author"
@@ -496,8 +497,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "decidim_commentable_type", null: false
     t.integer "decidim_commentable_id", null: false
     t.integer "decidim_author_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "depth", default: 0, null: false
     t.integer "alignment", default: 0, null: false
     t.integer "decidim_user_group_id"
@@ -508,7 +509,7 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "comments_count", default: 0, null: false
     t.string "decidim_participatory_space_type"
     t.integer "decidim_participatory_space_id"
-    t.datetime "deleted_at"
+    t.datetime "deleted_at", precision: nil
     t.integer "up_votes_count", default: 0, null: false
     t.integer "down_votes_count", default: 0, null: false
     t.index ["created_at"], name: "index_decidim_comments_comments_on_created_at"
@@ -527,110 +528,11 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "settings", default: {}
     t.integer "weight", default: 0
     t.jsonb "permissions"
-    t.datetime "published_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "published_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "participatory_space_type", null: false
     t.index ["participatory_space_id", "participatory_space_type"], name: "index_decidim_components_on_decidim_participatory_space"
-  end
-
-  create_table "decidim_consultations", force: :cascade do |t|
-    t.string "slug", null: false
-    t.integer "decidim_organization_id"
-    t.jsonb "title", null: false
-    t.jsonb "subtitle", null: false
-    t.jsonb "description", null: false
-    t.string "banner_image"
-    t.string "introductory_video_url"
-    t.date "start_voting_date", null: false
-    t.integer "decidim_highlighted_scope_id"
-    t.datetime "published_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.date "end_voting_date", null: false
-    t.date "results_published_at"
-    t.string "introductory_image"
-    t.index ["decidim_highlighted_scope_id"], name: "index_decidim_consultations_on_decidim_highlighted_scope_id"
-    t.index ["decidim_organization_id", "slug"], name: "index_unique_consultation_slug_and_organization", unique: true
-    t.index ["decidim_organization_id"], name: "index_decidim_consultations_on_decidim_organization_id"
-    t.index ["description"], name: "decidim_consultations_description_search"
-    t.index ["published_at"], name: "index_decidim_consultations_on_published_at"
-    t.index ["subtitle"], name: "decidim_consultations_subtitle_search"
-    t.index ["title"], name: "decidim_consultations_title_search"
-  end
-
-  create_table "decidim_consultations_questions", force: :cascade do |t|
-    t.bigint "decidim_consultation_id"
-    t.bigint "decidim_scope_id"
-    t.jsonb "title", null: false
-    t.jsonb "subtitle", null: false
-    t.jsonb "what_is_decided", null: false
-    t.jsonb "promoter_group", null: false
-    t.jsonb "participatory_scope", null: false
-    t.jsonb "question_context"
-    t.string "banner_image"
-    t.string "reference"
-    t.string "hashtag"
-    t.datetime "published_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.integer "decidim_organization_id", null: false
-    t.string "slug", null: false
-    t.integer "votes_count", default: 0, null: false
-    t.jsonb "origin_scope"
-    t.jsonb "origin_title"
-    t.string "origin_url"
-    t.string "i_frame_url"
-    t.boolean "external_voting"
-    t.integer "responses_count", default: 0, null: false
-    t.string "hero_image"
-    t.integer "order"
-    t.integer "max_votes"
-    t.integer "min_votes"
-    t.integer "response_groups_count", default: 0, null: false
-    t.jsonb "instructions"
-    t.integer "comments_count", default: 0, null: false
-    t.integer "follows_count", default: 0, null: false
-    t.index ["decidim_consultation_id"], name: "index_consultations_questions_on_consultation_id"
-    t.index ["decidim_organization_id", "slug"], name: "index_unique_question_slug_and_organization", unique: true
-    t.index ["decidim_scope_id"], name: "index_decidim_consultations_questions_on_decidim_scope_id"
-    t.index ["origin_scope"], name: "consultation_questions_origin_scope_search"
-    t.index ["origin_title"], name: "consultation_questions_origin_title_search"
-    t.index ["published_at"], name: "index_decidim_consultations_questions_on_published_at"
-  end
-
-  create_table "decidim_consultations_response_groups", force: :cascade do |t|
-    t.jsonb "title"
-    t.bigint "decidim_consultations_questions_id"
-    t.integer "responses_count", default: 0, null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.index ["decidim_consultations_questions_id"], name: "index_consultations_response_groups_on_consultation_questions"
-  end
-
-  create_table "decidim_consultations_responses", force: :cascade do |t|
-    t.jsonb "title"
-    t.bigint "decidim_consultations_questions_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.integer "votes_count", default: 0, null: false
-    t.bigint "decidim_consultations_response_group_id"
-    t.index ["decidim_consultations_questions_id"], name: "index_consultations_responses_on_consultation_questions"
-    t.index ["decidim_consultations_response_group_id"], name: "index_consultations_response_groups_on_consultation_responses"
-  end
-
-  create_table "decidim_consultations_votes", force: :cascade do |t|
-    t.bigint "decidim_consultation_question_id"
-    t.bigint "decidim_author_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.integer "decidim_user_group_id"
-    t.bigint "decidim_consultations_response_id"
-    t.index ["decidim_author_id"], name: "index_consultations_votes_on_author"
-    t.index ["decidim_consultation_question_id", "decidim_author_id", "decidim_user_group_id"], name: "index_question_votes_author_unique", unique: true
-    t.index ["decidim_consultation_question_id"], name: "index_consultations_votes_on_consultation_question"
-    t.index ["decidim_consultations_response_id"], name: "index_consultations_votes_on_consultations_response_id"
-    t.index ["decidim_user_group_id"], name: "index_decidim_consultations_votes_on_decidim_user_group_id"
   end
 
   create_table "decidim_content_block_attachments", force: :cascade do |t|
@@ -644,12 +546,12 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "manifest_name", null: false
     t.string "scope_name", null: false
     t.jsonb "settings"
-    t.datetime "published_at"
+    t.datetime "published_at", precision: nil
     t.integer "weight"
     t.jsonb "images", default: {}
     t.integer "scoped_resource_id"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["decidim_organization_id", "scope_name", "scoped_resource_id", "manifest_name"], name: "idx_decidim_content_blocks_org_id_scope_scope_id_manifest"
     t.index ["decidim_organization_id"], name: "index_decidim_content_blocks_on_decidim_organization_id"
     t.index ["manifest_name"], name: "index_decidim_content_blocks_on_manifest_name"
@@ -669,22 +571,22 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "title"
     t.jsonb "description"
     t.jsonb "instructions"
-    t.datetime "start_time"
-    t.datetime "end_time"
+    t.datetime "start_time", precision: nil
+    t.datetime "end_time", precision: nil
     t.string "image"
     t.integer "decidim_component_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "information_updates"
     t.integer "decidim_author_id", null: false
     t.string "reference"
     t.integer "decidim_user_group_id"
     t.string "decidim_author_type", null: false
-    t.datetime "closed_at"
+    t.datetime "closed_at", precision: nil
     t.jsonb "conclusions"
     t.integer "endorsements_count", default: 0, null: false
     t.integer "comments_count", default: 0, null: false
-    t.datetime "last_comment_at"
+    t.datetime "last_comment_at", precision: nil
     t.integer "last_comment_by_id"
     t.string "last_comment_by_type"
     t.bigint "decidim_scope_id"
@@ -701,8 +603,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_editor_images", force: :cascade do |t|
     t.bigint "decidim_author_id", null: false
     t.bigint "decidim_organization_id", null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["decidim_author_id"], name: "decidim_editor_images_author"
     t.index ["decidim_organization_id"], name: "decidim_editor_images_constraint_organization"
   end
@@ -713,8 +615,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "decidim_author_type"
     t.bigint "decidim_author_id"
     t.integer "decidim_user_group_id", default: 0
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_author_type", "decidim_author_id"], name: "idx_endorsements_authors"
     t.index ["decidim_user_group_id"], name: "index_decidim_endorsements_on_decidim_user_group_id"
     t.index ["resource_type", "resource_id", "decidim_author_type", "decidim_author_id", "decidim_user_group_id"], name: "idx_endorsements_rsrcs_and_authors", unique: true
@@ -725,8 +627,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "decidim_user_id", null: false
     t.string "decidim_followable_type"
     t.bigint "decidim_followable_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_followable_id", "decidim_followable_type"], name: "index_follows_followable_id_and_type"
     t.index ["decidim_user_id", "decidim_followable_id", "decidim_followable_type"], name: "index_uniq_on_follows_user_and_followable", unique: true
     t.index ["decidim_user_id"], name: "index_decidim_follows_on_decidim_user_id"
@@ -755,8 +657,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "decidim_user_id"
     t.integer "decidim_questionnaire_id"
     t.integer "decidim_question_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "session_token", default: "", null: false
     t.string "ip_hash"
     t.index ["decidim_question_id"], name: "index_decidim_forms_answers_question_id"
@@ -773,8 +675,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "condition_type", default: 0, null: false
     t.jsonb "condition_value"
     t.boolean "mandatory", default: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_answer_option_id"], name: "decidim_forms_display_condition_answer_option"
     t.index ["decidim_condition_question_id"], name: "decidim_forms_display_condition_condition_question"
     t.index ["decidim_question_id"], name: "decidim_forms_display_condition_question"
@@ -794,9 +696,9 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "tos"
     t.string "questionnaire_for_type"
     t.integer "questionnaire_for_id"
-    t.datetime "published_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "published_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "salt"
     t.index ["questionnaire_for_type", "questionnaire_for_id"], name: "index_decidim_forms_questionnaires_questionnaire_for"
   end
@@ -809,9 +711,13 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "body"
     t.jsonb "description"
     t.integer "max_choices"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "max_characters", default: 0
+    t.integer "answer_options_count", default: 0, null: false
+    t.integer "matrix_rows_count", default: 0, null: false
+    t.integer "display_conditions_count", default: 0, null: false
+    t.integer "display_conditions_for_other_questions_count", default: 0, null: false
     t.index ["decidim_questionnaire_id"], name: "index_decidim_forms_questions_on_decidim_questionnaire_id"
     t.index ["position"], name: "index_decidim_forms_questions_on_position"
   end
@@ -826,8 +732,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_hashtags", force: :cascade do |t|
     t.bigint "decidim_organization_id"
     t.string "name"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_organization_id"], name: "index_decidim_hashtags_on_decidim_organization_id"
     t.index ["name"], name: "index_decidim_hashtags_on_name"
   end
@@ -836,10 +742,9 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "provider", null: false
     t.string "uid", null: false
     t.integer "decidim_user_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "decidim_organization_id"
-    t.string "rrn_hash"
     t.index ["decidim_organization_id"], name: "index_decidim_identities_on_decidim_organization_id"
     t.index ["decidim_user_id"], name: "index_decidim_identities_on_decidim_user_id"
     t.index ["provider", "uid", "decidim_organization_id"], name: "decidim_identities_provider_uid_organization_unique", unique: true
@@ -848,11 +753,11 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_impersonation_logs", force: :cascade do |t|
     t.bigint "decidim_admin_id"
     t.bigint "decidim_user_id"
-    t.datetime "started_at"
-    t.datetime "ended_at"
-    t.datetime "expired_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "started_at", precision: nil
+    t.datetime "ended_at", precision: nil
+    t.datetime "expired_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.text "reason"
     t.index ["decidim_admin_id"], name: "index_decidim_impersonation_logs_on_decidim_admin_id"
     t.index ["decidim_user_id"], name: "index_decidim_impersonation_logs_on_decidim_user_id"
@@ -865,8 +770,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "duration"
     t.jsonb "title"
     t.jsonb "description"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_agenda_id"], name: "index_decidim_meetings_agenda_items_on_decidim_agenda_id"
     t.index ["parent_id"], name: "index_decidim_meetings_agenda_items_on_parent_id"
   end
@@ -875,8 +780,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "title"
     t.bigint "decidim_meeting_id", null: false
     t.boolean "visible"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_meeting_id"], name: "index_decidim_meetings_agendas_on_decidim_meeting_id"
   end
 
@@ -900,8 +805,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "decidim_user_id"
     t.bigint "decidim_questionnaire_id"
     t.bigint "decidim_question_id"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["decidim_question_id"], name: "index_decidim_meetings_answers_question_id"
     t.index ["decidim_questionnaire_id"], name: "index_decidim_meetings_answers_on_decidim_questionnaire_id"
     t.index ["decidim_user_id"], name: "index_decidim_meetings_answers_on_decidim_user_id"
@@ -910,11 +815,11 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_meetings_invites", force: :cascade do |t|
     t.bigint "decidim_user_id", null: false
     t.bigint "decidim_meeting_id", null: false
-    t.datetime "sent_at"
-    t.datetime "accepted_at"
-    t.datetime "rejected_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "sent_at", precision: nil
+    t.datetime "accepted_at", precision: nil
+    t.datetime "rejected_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_meeting_id"], name: "index_decidim_meetings_invites_on_decidim_meeting_id"
     t.index ["decidim_user_id"], name: "index_decidim_meetings_invites_on_decidim_user_id"
   end
@@ -922,16 +827,16 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_meetings_meetings", id: :serial, force: :cascade do |t|
     t.jsonb "title"
     t.jsonb "description"
-    t.datetime "start_time"
-    t.datetime "end_time"
+    t.datetime "start_time", precision: nil
+    t.datetime "end_time", precision: nil
     t.text "address"
     t.jsonb "location"
     t.jsonb "location_hints"
     t.integer "decidim_component_id"
     t.integer "decidim_author_id"
     t.integer "decidim_scope_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "closing_report"
     t.integer "attendees_count"
     t.integer "contributions_count"
@@ -951,23 +856,24 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "decidim_user_group_id"
     t.integer "comments_count", default: 0, null: false
     t.string "online_meeting_url"
-    t.string "type_of_meeting", default: "in_person"
-    t.string "registration_type", default: "registration_disabled", null: false
     t.string "registration_url"
     t.string "salt"
-    t.integer "follows_count", default: 0, null: false
     t.boolean "customize_registration_email", default: false
     t.jsonb "registration_email_custom_content"
-    t.datetime "published_at"
+    t.integer "follows_count", default: 0, null: false
+    t.datetime "published_at", precision: nil
     t.string "video_url"
     t.string "audio_url"
     t.boolean "closing_visible"
     t.boolean "comments_enabled", default: true
-    t.datetime "comments_start_time"
-    t.datetime "comments_end_time"
+    t.datetime "comments_start_time", precision: nil
+    t.datetime "comments_end_time", precision: nil
     t.string "state"
     t.integer "iframe_access_level", default: 0
     t.integer "iframe_embed_type", default: 0
+    t.integer "type_of_meeting", default: 0, null: false
+    t.integer "registration_type", default: 0, null: false
+    t.datetime "withdrawn_at", precision: nil
     t.index ["decidim_author_id", "decidim_author_type"], name: "index_decidim_meetings_meetings_on_author"
     t.index ["decidim_author_id"], name: "index_decidim_meetings_meetings_on_decidim_author_id"
     t.index ["decidim_component_id"], name: "index_decidim_meetings_meetings_on_decidim_component_id"
@@ -976,16 +882,16 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
 
   create_table "decidim_meetings_polls", force: :cascade do |t|
     t.bigint "decidim_meeting_id"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["decidim_meeting_id"], name: "index_decidim_meetings_polls_on_decidim_meeting_id"
   end
 
   create_table "decidim_meetings_questionnaires", force: :cascade do |t|
     t.string "questionnaire_for_type"
     t.bigint "questionnaire_for_id"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["questionnaire_for_type", "questionnaire_for_id"], name: "index_decidim_meetings_questionnaires_questionnaire_for"
   end
 
@@ -995,8 +901,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "question_type"
     t.jsonb "body"
     t.integer "max_choices"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.integer "status", default: 0
     t.index ["decidim_questionnaire_id"], name: "index_decidim_meetings_questions_on_decidim_questionnaire_id"
     t.index ["position"], name: "index_decidim_meetings_questions_on_position"
@@ -1005,10 +911,10 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_meetings_registrations", force: :cascade do |t|
     t.bigint "decidim_user_id", null: false
     t.bigint "decidim_meeting_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "code"
-    t.datetime "validated_at"
+    t.datetime "validated_at", precision: nil
     t.bigint "decidim_user_group_id"
     t.boolean "public_participation", default: false
     t.index ["decidim_meeting_id"], name: "index_decidim_meetings_registrations_on_decidim_meeting_id"
@@ -1021,22 +927,22 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "title"
     t.jsonb "description"
     t.bigint "decidim_meeting_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_meeting_id"], name: "index_decidim_meetings_services_on_decidim_meeting_id"
   end
 
   create_table "decidim_messaging_conversations", force: :cascade do |t|
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
   end
 
   create_table "decidim_messaging_messages", force: :cascade do |t|
     t.bigint "decidim_conversation_id", null: false
     t.bigint "decidim_sender_id", null: false
     t.text "body", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_conversation_id"], name: "index_decidim_messaging_messages_on_decidim_conversation_id"
     t.index ["decidim_sender_id"], name: "index_decidim_messaging_messages_on_decidim_sender_id"
   end
@@ -1044,8 +950,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_messaging_participations", force: :cascade do |t|
     t.bigint "decidim_conversation_id", null: false
     t.bigint "decidim_participant_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_conversation_id"], name: "index_conversation_participations_on_conversation_id"
     t.index ["decidim_participant_id"], name: "index_conversation_participations_on_participant_id"
   end
@@ -1053,9 +959,9 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_messaging_receipts", force: :cascade do |t|
     t.bigint "decidim_message_id", null: false
     t.bigint "decidim_recipient_id", null: false
-    t.datetime "read_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "read_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_message_id"], name: "index_decidim_messaging_receipts_on_decidim_message_id"
     t.index ["decidim_recipient_id"], name: "index_decidim_messaging_receipts_on_decidim_recipient_id"
   end
@@ -1085,9 +991,9 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "decidim_reportable_type", null: false
     t.integer "decidim_reportable_id", null: false
     t.integer "report_count", default: 0, null: false
-    t.datetime "hidden_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "hidden_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "decidim_participatory_space_type", null: false
     t.text "reported_content"
     t.index ["decidim_participatory_space_id", "decidim_participatory_space_type"], name: "decidim_moderations_participatory_space"
@@ -1102,9 +1008,9 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "author_id"
     t.integer "total_recipients"
     t.integer "total_deliveries"
-    t.datetime "sent_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "sent_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "extended_data", default: {}
     t.index ["author_id"], name: "index_decidim_newsletters_on_author_id"
     t.index ["organization_id"], name: "index_decidim_newsletters_on_organization_id"
@@ -1116,20 +1022,19 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "decidim_resource_id", null: false
     t.string "event_name", null: false
     t.string "event_class", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "extra"
     t.index ["decidim_resource_id"], name: "index_decidim_notifications_on_decidim_resource_id"
     t.index ["decidim_user_id"], name: "index_decidim_notifications_on_decidim_user_id"
   end
 
   create_table "decidim_organizations", id: :serial, force: :cascade do |t|
-    t.string "name", null: false
     t.string "host", null: false
     t.string "default_locale", null: false
     t.string "available_locales", default: [], array: true
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "description"
     t.string "logo"
     t.string "twitter_handler"
@@ -1138,7 +1043,6 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "facebook_handler"
     t.string "youtube_handler"
     t.string "github_handler"
-    t.string "official_img_header"
     t.string "official_img_footer"
     t.string "official_url"
     t.string "reference_prefix", null: false
@@ -1158,7 +1062,7 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "highlighted_content_banner_action_subtitle"
     t.string "highlighted_content_banner_action_url"
     t.string "highlighted_content_banner_image"
-    t.datetime "tos_version"
+    t.datetime "tos_version", precision: nil
     t.boolean "badges_enabled", default: false, null: false
     t.boolean "send_welcome_notification", default: false, null: false
     t.jsonb "welcome_notification_subject"
@@ -1172,23 +1076,24 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.boolean "force_users_to_authenticate_before_access_organization", default: false
     t.jsonb "omniauth_settings"
     t.boolean "rich_text_editor_in_public_views", default: false
-    t.jsonb "admin_terms_of_use_body"
+    t.jsonb "admin_terms_of_service_body"
     t.string "time_zone", limit: 255, default: "UTC"
     t.boolean "enable_machine_translations", default: false
     t.integer "comments_max_length", default: 1000
     t.jsonb "file_upload_settings"
     t.string "machine_translation_display_priority", default: "original", null: false
-    t.string "external_domain_whitelist", default: [], array: true
+    t.string "external_domain_allowlist", default: [], array: true
     t.boolean "enable_participatory_space_filters", default: true
+    t.jsonb "content_security_policy", default: {}
+    t.jsonb "name", default: {}, null: false
     t.index ["host"], name: "index_decidim_organizations_on_host", unique: true
-    t.index ["name"], name: "index_decidim_organizations_on_name", unique: true
   end
 
   create_table "decidim_pages_pages", id: :serial, force: :cascade do |t|
     t.jsonb "body"
     t.integer "decidim_component_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_component_id"], name: "index_decidim_pages_pages_on_decidim_component_id"
   end
 
@@ -1197,8 +1102,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "description", null: false
     t.string "hero_image"
     t.integer "decidim_organization_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "hashtag"
     t.string "group_url"
     t.jsonb "developer_group"
@@ -1214,11 +1119,11 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_participatory_process_steps", id: :serial, force: :cascade do |t|
     t.jsonb "title", null: false
     t.jsonb "description"
-    t.date "start_date"
-    t.date "end_date"
+    t.datetime "start_date", precision: nil
+    t.datetime "end_date", precision: nil
     t.integer "decidim_participatory_process_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.boolean "active", default: false
     t.integer "position"
     t.jsonb "cta_text", default: {}
@@ -1232,8 +1137,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_participatory_process_types", force: :cascade do |t|
     t.jsonb "title", null: false
     t.bigint "decidim_organization_id"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["decidim_organization_id"], name: "index_decidim_process_types_on_decidim_organization_id"
   end
 
@@ -1241,8 +1146,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "decidim_user_id"
     t.integer "decidim_participatory_process_id"
     t.string "role"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_participatory_process_id", "decidim_user_id", "role"], name: "index_unique_user_and_process_role", unique: true
     t.index ["decidim_user_id"], name: "idx_proces_user_role_on_user_id"
   end
@@ -1251,8 +1156,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "slug", null: false
     t.string "hashtag"
     t.integer "decidim_organization_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "title", null: false
     t.jsonb "subtitle", null: false
     t.jsonb "short_description", null: false
@@ -1260,7 +1165,7 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "hero_image"
     t.string "banner_image"
     t.boolean "promoted", default: false
-    t.datetime "published_at"
+    t.datetime "published_at", precision: nil
     t.jsonb "developer_group"
     t.date "end_date"
     t.jsonb "meta_scope"
@@ -1307,8 +1212,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "decidim_user_id"
     t.integer "privatable_to_id"
     t.string "privatable_to_type"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_user_id"], name: "index_decidim_spaces_users_on_private_user_id"
     t.index ["privatable_to_type", "privatable_to_id"], name: "space_privatable_to_privatable_id"
   end
@@ -1316,8 +1221,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_proposals_collaborative_draft_collaborator_requests", force: :cascade do |t|
     t.bigint "decidim_proposals_collaborative_draft_id", null: false
     t.bigint "decidim_user_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_proposals_collaborative_draft_id"], name: "index_collab_requests_on_decidim_proposals_collab_draft_id"
     t.index ["decidim_user_id"], name: "index_collab_requests_on_decidim_user_id"
   end
@@ -1327,24 +1232,23 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.text "body", null: false
     t.integer "decidim_component_id", null: false
     t.integer "decidim_scope_id"
-    t.string "state"
     t.string "reference"
     t.text "address"
     t.float "latitude"
     t.float "longitude"
-    t.datetime "published_at"
+    t.datetime "published_at", precision: nil
     t.integer "authors_count", default: 0, null: false
     t.integer "versions_count", default: 0, null: false
     t.integer "contributions_count", default: 0, null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "coauthorships_count", default: 0, null: false
     t.integer "comments_count", default: 0, null: false
     t.integer "follows_count", default: 0, null: false
+    t.integer "state", default: 0, null: false
     t.index ["body"], name: "decidim_proposals_collaborative_draft_body_search"
     t.index ["decidim_component_id"], name: "decidim_proposals_collaborative_drafts_on_decidim_component_id"
     t.index ["decidim_scope_id"], name: "decidim_proposals_collaborative_drafts_on_decidim_scope_id"
-    t.index ["state"], name: "decidim_proposals_collaborative_drafts_on_state"
     t.index ["title"], name: "decidim_proposals_collaborative_drafts_title_search"
     t.index ["updated_at"], name: "decidim_proposals_collaborative_drafts_on_updated_at"
   end
@@ -1353,8 +1257,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.jsonb "title"
     t.jsonb "description"
     t.bigint "decidim_component_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_component_id"], name: "idx_participatory_texts_on_decidim_component_id"
   end
 
@@ -1362,18 +1266,29 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "decidim_proposal_id", null: false
     t.bigint "decidim_author_id", null: false
     t.text "body", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["created_at"], name: "index_decidim_proposals_proposal_notes_on_created_at"
     t.index ["decidim_author_id"], name: "decidim_proposals_proposal_note_author"
     t.index ["decidim_proposal_id"], name: "decidim_proposals_proposal_note_proposal"
   end
 
+  create_table "decidim_proposals_proposal_states", force: :cascade do |t|
+    t.jsonb "title"
+    t.jsonb "announcement_title"
+    t.string "token", null: false
+    t.bigint "decidim_component_id", null: false
+    t.integer "proposals_count", default: 0, null: false
+    t.string "bg_color", default: "#F6F8FA", null: false
+    t.string "text_color", default: "#4B5058", null: false
+    t.index ["decidim_component_id"], name: "index_decidim_proposals_proposal_states_on_decidim_component_id"
+  end
+
   create_table "decidim_proposals_proposal_votes", id: :serial, force: :cascade do |t|
     t.integer "decidim_proposal_id", null: false
     t.integer "decidim_author_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.boolean "temporary", default: false, null: false
     t.index ["decidim_author_id"], name: "decidim_proposals_proposal_vote_author"
     t.index ["decidim_proposal_id", "decidim_author_id"], name: "decidim_proposals_proposal_vote_proposal_author_unique", unique: true
@@ -1383,17 +1298,16 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_proposals_proposals", id: :serial, force: :cascade do |t|
     t.integer "decidim_component_id", null: false
     t.integer "decidim_scope_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "proposal_votes_count", default: 0, null: false
-    t.string "state"
-    t.datetime "answered_at"
+    t.datetime "answered_at", precision: nil
     t.jsonb "answer"
     t.string "reference"
     t.text "address"
     t.float "latitude"
     t.float "longitude"
-    t.datetime "published_at"
+    t.datetime "published_at", precision: nil
     t.integer "proposal_notes_count", default: 0, null: false
     t.integer "coauthorships_count", default: 0, null: false
     t.string "participatory_text_level"
@@ -1403,11 +1317,14 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.decimal "cost"
     t.jsonb "cost_report"
     t.jsonb "execution_period"
-    t.datetime "state_published_at"
+    t.datetime "state_published_at", precision: nil
     t.jsonb "title"
     t.jsonb "body"
     t.integer "comments_count", default: 0, null: false
     t.integer "follows_count", default: 0, null: false
+    t.integer "old_state", default: 0, null: false
+    t.datetime "withdrawn_at", precision: nil
+    t.integer "decidim_proposals_proposal_state_id"
     t.integer "valuation_assignments_count", default: 0
     t.index "md5((body)::text)", name: "decidim_proposals_proposal_body_search"
     t.index "md5((title)::text)", name: "decidim_proposals_proposal_title_search"
@@ -1415,43 +1332,41 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.index ["decidim_component_id"], name: "index_decidim_proposals_proposals_on_decidim_component_id"
     t.index ["decidim_scope_id"], name: "index_decidim_proposals_proposals_on_decidim_scope_id"
     t.index ["proposal_votes_count"], name: "index_decidim_proposals_proposals_on_proposal_votes_count"
-    t.index ["state"], name: "index_decidim_proposals_proposals_on_state"
   end
 
   create_table "decidim_proposals_valuation_assignments", force: :cascade do |t|
     t.bigint "decidim_proposal_id", null: false
     t.string "valuator_role_type", null: false
     t.bigint "valuator_role_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_proposal_id"], name: "decidim_proposals_valuation_assignment_proposal"
     t.index ["valuator_role_type", "valuator_role_id"], name: "decidim_proposals_valuation_assignment_valuator_role"
   end
 
   create_table "decidim_reminder_deliveries", force: :cascade do |t|
     t.bigint "decidim_reminder_id"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["decidim_reminder_id"], name: "index_decidim_reminder_deliveries_on_decidim_reminder_id"
   end
 
   create_table "decidim_reminder_records", force: :cascade do |t|
-    t.string "state", default: "active"
     t.string "string", default: "active"
     t.bigint "decidim_reminder_id"
     t.string "remindable_type", null: false
     t.bigint "remindable_id", null: false
+    t.integer "state", default: 0, null: false
     t.index ["decidim_reminder_id"], name: "index_decidim_reminder_records_on_decidim_reminder_id"
     t.index ["remindable_type", "remindable_id"], name: "index_decidim_reminder_records_remindable"
-    t.index ["state"], name: "index_decidim_reminder_records_on_state"
     t.index ["string"], name: "index_decidim_reminder_records_on_string"
   end
 
   create_table "decidim_reminders", force: :cascade do |t|
     t.bigint "decidim_user_id", null: false
     t.bigint "decidim_component_id"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["decidim_component_id"], name: "index_decidim_reminders_on_decidim_component_id"
     t.index ["decidim_user_id"], name: "index_decidim_reminders_on_decidim_user_id"
   end
@@ -1461,8 +1376,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "decidim_user_id", null: false
     t.string "reason", null: false
     t.text "details"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "locale"
     t.index ["decidim_moderation_id", "decidim_user_id"], name: "decidim_reports_moderation_user_unique", unique: true
     t.index ["decidim_moderation_id"], name: "decidim_reports_moderation"
@@ -1485,8 +1400,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "resource_type"
     t.bigint "resource_id"
     t.jsonb "permissions", default: {}
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["resource_type", "resource_id"], name: "index_decidim_resource_permissions_on_r_type_and_r_id", unique: true
   end
 
@@ -1499,8 +1414,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
 
   create_table "decidim_scopes", id: :serial, force: :cascade do |t|
     t.integer "decidim_organization_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "name", null: false
     t.integer "scope_type_id"
     t.integer "parent_id"
@@ -1519,15 +1434,15 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.text "content_c"
     t.text "content_d"
     t.string "locale", null: false
-    t.datetime "datetime"
+    t.datetime "datetime", precision: nil
     t.bigint "decidim_scope_id"
     t.string "decidim_participatory_space_type"
     t.bigint "decidim_participatory_space_id"
     t.bigint "decidim_organization_id"
     t.string "resource_type"
     t.bigint "resource_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_organization_id"], name: "index_decidim_searchable_resources_on_decidim_organization_id"
     t.index ["decidim_participatory_space_type", "decidim_participatory_space_id"], name: "index_decidim_searchable_resource_on_pspace_type_and_pspace_id"
     t.index ["decidim_scope_id"], name: "index_decidim_searchable_resources_on_decidim_scope_id"
@@ -1541,9 +1456,9 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "token_for_id", null: false
     t.string "token", null: false
     t.integer "times_used", default: 0
-    t.datetime "created_at"
-    t.datetime "last_used_at"
-    t.datetime "expires_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "last_used_at", precision: nil
+    t.datetime "expires_at", precision: nil
     t.index ["decidim_organization_id"], name: "index_decidim_share_tokens_on_decidim_organization_id"
     t.index ["decidim_user_id"], name: "index_decidim_share_tokens_on_decidim_user_id"
     t.index ["token_for_type", "token_for_id"], name: "decidim_share_tokens_token_for"
@@ -1557,8 +1472,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "mounted_engine_name"
     t.string "route_name"
     t.jsonb "params"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["decidim_organization_id", "identifier"], name: "idx_decidim_short_links_organization_id_identifier", unique: true
     t.index ["decidim_organization_id"], name: "index_decidim_short_links_on_decidim_organization_id"
     t.index ["mounted_engine_name"], name: "index_decidim_short_links_on_mounted_engine_name"
@@ -1571,17 +1486,17 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "decidim_proposals_component_id"
     t.integer "dice", null: false
     t.integer "target_items", null: false
-    t.datetime "request_timestamp", null: false
+    t.datetime "request_timestamp", precision: nil, null: false
     t.jsonb "selected_proposals"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.jsonb "witnesses"
     t.jsonb "additional_info"
     t.bigint "decidim_author_id", null: false
     t.string "reference"
     t.jsonb "title"
     t.jsonb "cancel_reason"
-    t.datetime "cancelled_on"
+    t.datetime "cancelled_on", precision: nil
     t.integer "cancelled_by_user_id"
     t.jsonb "candidate_proposals"
     t.string "decidim_author_type", null: false
@@ -1607,10 +1522,9 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "slug", null: false
     t.jsonb "content", null: false
     t.integer "decidim_organization_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "weight"
-    t.boolean "show_in_footer", default: false, null: false
     t.bigint "topic_id"
     t.boolean "allow_public_access", default: false, null: false
     t.index ["decidim_organization_id"], name: "index_decidim_static_pages_on_decidim_organization_id"
@@ -1619,8 +1533,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
 
   create_table "decidim_surveys_surveys", id: :serial, force: :cascade do |t|
     t.integer "decidim_component_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_component_id"], name: "index_decidim_surveys_surveys_on_decidim_component_id"
   end
 
@@ -1628,13 +1542,13 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
     t.string "reset_password_token"
-    t.datetime "reset_password_sent_at"
-    t.datetime "remember_created_at"
+    t.datetime "reset_password_sent_at", precision: nil
+    t.datetime "remember_created_at", precision: nil
     t.integer "failed_attempts", default: 0, null: false
     t.string "unlock_token"
-    t.datetime "locked_at"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "locked_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["email"], name: "index_decidim_system_admins_on_email", unique: true
     t.index ["reset_password_token"], name: "index_decidim_system_admins_on_reset_password_token", unique: true
   end
@@ -1645,8 +1559,10 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "templatable_id"
     t.jsonb "name", null: false
     t.jsonb "description"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
+    t.json "field_values", default: {}
+    t.string "target"
     t.index ["decidim_organization_id"], name: "index_decidim_templates_organization"
     t.index ["templatable_type", "templatable_id"], name: "index_decidim_templates_templatable"
   end
@@ -1677,16 +1593,16 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.bigint "decidim_user_id"
     t.integer "blocking_user_id"
     t.text "justification"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_user_id"], name: "index_decidim_user_blocks_on_decidim_user_id"
   end
 
   create_table "decidim_user_group_memberships", id: :serial, force: :cascade do |t|
     t.integer "decidim_user_id", null: false
     t.integer "decidim_user_group_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "role", default: "requested", null: false
     t.index ["decidim_user_group_id", "decidim_user_id"], name: "index_user_group_memberships_group_id_user_id"
     t.index ["decidim_user_group_id"], name: "index_decidim_user_group_memberships_on_decidim_user_group_id"
@@ -1698,8 +1614,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_user_moderations", force: :cascade do |t|
     t.bigint "decidim_user_id"
     t.integer "report_count", default: 0, null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_user_id"], name: "index_decidim_user_moderations_on_decidim_user_id"
   end
 
@@ -1708,50 +1624,50 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "user_id", null: false
     t.string "reason"
     t.text "details"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
   end
 
   create_table "decidim_users", id: :serial, force: :cascade do |t|
-    t.string "email", default: ""
+    t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
     t.string "reset_password_token"
-    t.datetime "reset_password_sent_at"
-    t.datetime "remember_created_at"
+    t.datetime "reset_password_sent_at", precision: nil
+    t.datetime "remember_created_at", precision: nil
     t.integer "sign_in_count", default: 0, null: false
-    t.datetime "current_sign_in_at"
-    t.datetime "last_sign_in_at"
+    t.datetime "current_sign_in_at", precision: nil
+    t.datetime "last_sign_in_at", precision: nil
     t.string "current_sign_in_ip"
     t.string "last_sign_in_ip"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "invitation_token"
-    t.datetime "invitation_created_at"
-    t.datetime "invitation_sent_at"
-    t.datetime "invitation_accepted_at"
+    t.datetime "invitation_created_at", precision: nil
+    t.datetime "invitation_sent_at", precision: nil
+    t.datetime "invitation_accepted_at", precision: nil
     t.integer "invitation_limit"
     t.string "invited_by_type"
     t.integer "invited_by_id"
     t.integer "invitations_count", default: 0
     t.integer "decidim_organization_id"
     t.string "confirmation_token"
-    t.datetime "confirmed_at"
-    t.datetime "confirmation_sent_at"
+    t.datetime "confirmed_at", precision: nil
+    t.datetime "confirmation_sent_at", precision: nil
     t.string "unconfirmed_email"
     t.string "name", null: false
     t.string "locale"
     t.string "avatar"
     t.text "delete_reason"
-    t.datetime "deleted_at"
+    t.datetime "deleted_at", precision: nil
     t.boolean "admin", default: false, null: false
     t.boolean "managed", default: false, null: false
     t.string "roles", default: [], array: true
     t.string "nickname", limit: 20, default: "", null: false
     t.string "personal_url"
     t.text "about"
-    t.datetime "accepted_tos_version"
+    t.datetime "accepted_tos_version", precision: nil
     t.string "newsletter_token", default: ""
-    t.datetime "newsletter_notifications_at"
+    t.datetime "newsletter_notifications_at", precision: nil
     t.string "type", null: false
     t.jsonb "extended_data", default: {}
     t.integer "following_count", default: 0, null: false
@@ -1759,22 +1675,22 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "notification_types", default: "all", null: false
     t.integer "failed_attempts", default: 0, null: false
     t.string "unlock_token"
-    t.datetime "locked_at"
+    t.datetime "locked_at", precision: nil
     t.string "session_token"
     t.string "direct_message_types", default: "all", null: false
     t.boolean "blocked", default: false, null: false
-    t.datetime "blocked_at"
+    t.datetime "blocked_at", precision: nil
     t.integer "block_id"
     t.boolean "email_on_moderations", default: true
     t.integer "follows_count", default: 0, null: false
-    t.datetime "officialized_at"
-    t.jsonb "officialized_as"
-    t.datetime "admin_terms_accepted_at"
     t.jsonb "notification_settings", default: {}
     t.string "notifications_sending_frequency", default: "daily"
-    t.datetime "digest_sent_at"
-    t.datetime "password_updated_at"
+    t.datetime "digest_sent_at", precision: nil
+    t.datetime "password_updated_at", precision: nil
     t.string "previous_passwords", default: [], array: true
+    t.datetime "officialized_at", precision: nil
+    t.jsonb "officialized_as"
+    t.datetime "admin_terms_accepted_at", precision: nil
     t.index ["confirmation_token"], name: "index_decidim_users_on_confirmation_token", unique: true
     t.index ["decidim_organization_id"], name: "index_decidim_users_on_decidim_organization_id"
     t.index ["email", "decidim_organization_id"], name: "index_decidim_users_on_email_and_decidim_organization_id", unique: true, where: "((deleted_at IS NULL) AND (managed = false) AND ((type)::text = 'Decidim::User'::text))"
@@ -1796,8 +1712,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.integer "times", default: 0
     t.string "unique_id"
     t.boolean "solved", default: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["current_user_id"], name: "authorization_current_user"
     t.index ["managed_user_id"], name: "authorization_managed_user"
   end
@@ -1805,24 +1721,9 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   create_table "decidim_verifications_csv_data", force: :cascade do |t|
     t.string "email"
     t.bigint "decidim_organization_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["decidim_organization_id"], name: "index_verifications_csv_census_to_organization"
-  end
-
-  create_table "delayed_jobs", force: :cascade do |t|
-    t.integer "priority", default: 0, null: false
-    t.integer "attempts", default: 0, null: false
-    t.text "handler", null: false
-    t.text "last_error"
-    t.datetime "run_at"
-    t.datetime "locked_at"
-    t.datetime "failed_at"
-    t.string "locked_by"
-    t.string "queue"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-    t.index ["priority", "run_at"], name: "delayed_jobs_priority"
   end
 
   create_table "oauth_access_grants", force: :cascade do |t|
@@ -1831,8 +1732,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "token", null: false
     t.integer "expires_in", null: false
     t.text "redirect_uri", null: false
-    t.datetime "created_at", null: false
-    t.datetime "revoked_at"
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "revoked_at", precision: nil
     t.string "scopes"
     t.index ["application_id"], name: "index_oauth_access_grants_on_application_id"
     t.index ["resource_owner_id"], name: "index_oauth_access_grants_on_resource_owner_id"
@@ -1845,8 +1746,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "token", null: false
     t.string "refresh_token"
     t.integer "expires_in"
-    t.datetime "revoked_at"
-    t.datetime "created_at", null: false
+    t.datetime "revoked_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
     t.string "scopes"
     t.string "previous_refresh_token", default: "", null: false
     t.index ["application_id"], name: "index_oauth_access_tokens_on_application_id"
@@ -1865,8 +1766,8 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.text "redirect_uri", null: false
     t.string "scopes", default: "", null: false
     t.bigint "decidim_organization_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "type"
     t.boolean "confidential", default: true, null: false
     t.index ["decidim_organization_id"], name: "index_oauth_applications_on_decidim_organization_id"
@@ -1879,7 +1780,7 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
     t.string "event", null: false
     t.string "whodunnit"
     t.jsonb "object"
-    t.datetime "created_at"
+    t.datetime "created_at", precision: nil
     t.text "object_changes"
     t.index ["item_id", "item_type"], name: "index_versions_on_item_id_and_item_type"
     t.index ["item_type", "item_id"], name: "index_versions_on_item_type_and_item_id"
@@ -1887,13 +1788,16 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
 
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
   add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
-  add_foreign_key "decidim_anonymous_codes_groups", "decidim_organizations"
   add_foreign_key "decidim_area_types", "decidim_organizations"
   add_foreign_key "decidim_areas", "decidim_area_types", column: "area_type_id"
   add_foreign_key "decidim_areas", "decidim_organizations"
   add_foreign_key "decidim_assemblies", "decidim_assemblies_types"
   add_foreign_key "decidim_assemblies_settings", "decidim_organizations"
   add_foreign_key "decidim_attachments", "decidim_attachment_collections", column: "attachment_collection_id", name: "fk_decidim_attachments_attachment_collection_id", on_delete: :nullify
+  add_foreign_key "decidim_authorization_transfer_records", "decidim_authorization_transfers", column: "transfer_id"
+  add_foreign_key "decidim_authorization_transfers", "decidim_authorizations", column: "authorization_id"
+  add_foreign_key "decidim_authorization_transfers", "decidim_users", column: "source_user_id"
+  add_foreign_key "decidim_authorization_transfers", "decidim_users", column: "user_id"
   add_foreign_key "decidim_authorizations", "decidim_users"
   add_foreign_key "decidim_awesome_config_constraints", "decidim_awesome_config"
   add_foreign_key "decidim_awesome_editor_images", "decidim_organizations"
@@ -1902,10 +1806,6 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   add_foreign_key "decidim_budgets_orders", "decidim_budgets_budgets"
   add_foreign_key "decidim_budgets_projects", "decidim_budgets_budgets"
   add_foreign_key "decidim_categorizations", "decidim_categories"
-  add_foreign_key "decidim_consultations_response_groups", "decidim_consultations_questions", column: "decidim_consultations_questions_id"
-  add_foreign_key "decidim_consultations_responses", "decidim_consultations_questions", column: "decidim_consultations_questions_id"
-  add_foreign_key "decidim_consultations_responses", "decidim_consultations_response_groups"
-  add_foreign_key "decidim_consultations_votes", "decidim_consultations_responses"
   add_foreign_key "decidim_debates_debates", "decidim_scopes"
   add_foreign_key "decidim_editor_images", "decidim_organizations"
   add_foreign_key "decidim_editor_images", "decidim_users", column: "decidim_author_id"
@@ -1916,6 +1816,7 @@ ActiveRecord::Schema.define(version: 2024_09_23_110435) do
   add_foreign_key "decidim_participatory_processes", "decidim_organizations"
   add_foreign_key "decidim_participatory_processes", "decidim_participatory_process_types"
   add_foreign_key "decidim_participatory_processes", "decidim_scope_types"
+  add_foreign_key "decidim_proposals_proposals", "decidim_proposals_proposal_states"
   add_foreign_key "decidim_reminder_deliveries", "decidim_reminders"
   add_foreign_key "decidim_reminder_records", "decidim_reminders"
   add_foreign_key "decidim_reminders", "decidim_components"
