@@ -1,6 +1,6 @@
 const path = require("path");
 const miniCssExtractPlugin = require("mini-css-extract-plugin");
-const { inliningCss } = require("@rails/webpacker");
+const { inliningCss } = require("shakapacker");
 
 const overrideSassRule = (modifyConfig) => {
   const sassLoaderPath = path.resolve(__dirname, "loaders/decidim-sass-loader") // eslint-disable-line no-undef
@@ -24,6 +24,13 @@ const overrideSassRule = (modifyConfig) => {
       baseLoader = miniCssExtractPlugin.loader;
     }
 
+    // eslint-disable-next-line no-undef
+    let postCssConfig = path.resolve(__dirname, "../../../postcss.config.js");
+    if (postCssConfig.includes("node_modules")) {
+      // eslint-disable-next-line no-undef
+      postCssConfig = path.resolve(__dirname, "../../../../postcss.config.js");
+    }
+
     modifyConfig.module.rules.push({
       test: /\.(scss|sass)(\.erb)?$/i,
       use: [
@@ -37,13 +44,18 @@ const overrideSassRule = (modifyConfig) => {
         },
         {
           loader: "postcss-loader",
-          options: { sourceMap: true }
+          options: {
+            sourceMap: true,
+            postcssOptions: {
+              config: postCssConfig
+            }
+          }
         },
         {
           loader: sassLoaderPath
         }
       ]
-    })
+    });
   }
 
   return modifyConfig;
