@@ -14,7 +14,7 @@
 # the additional setup, and require it from the spec files that actually need
 # it.
 #
-# See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
+# See https://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
   # rspec-expectations config goes here. You can use an alternate
   # assertion/expectation library such as wrong or the stdlib/minitest
@@ -62,9 +62,7 @@ RSpec.configure do |config|
   #
   #   # Limits the available syntax to the non-monkey patched syntax that is
   #   # recommended. For more details, see:
-  #   #   - http://rspec.info/blog/2012/06/rspecs-new-expectation-syntax/
-  #   #   - http://www.teaisaweso.me/blog/2013/05/27/rspecs-new-message-expectation-syntax/
-  #   #   - http://rspec.info/blog/2014/05/notable-changes-in-rspec-3/#zero-monkey-patching-mode
+  #   # https://rspec.info/features/3-12/rspec-core/configuration/zero-monkey-patching-mode/
   #   config.disable_monkey_patching!
   #
   #   # Many RSpec users commonly either run the entire suite or an individual
@@ -93,16 +91,18 @@ RSpec.configure do |config|
   #   # test failures related to randomization by passing the same `--seed` value
   #   # as the one that triggered the failure.
   #   Kernel.srand config.seed
+
   config.order = :random
 
+  # Make :en locale available in tests.
   config.before do
-    I18n.available_locales = [:en, :es, :ca]
+    I18n.available_locales = [:eu, :en, :es]
     I18n.default_locale = :en
-    # rubocop:disable Rails/I18nLocaleAssignment
-    I18n.locale = :en
-    # rubocop:enable Rails/I18nLocaleAssignment
-    Decidim.available_locales = [:en, :es, :ca]
+    Decidim.available_locales = [:eu, :en, :es]
     Decidim.default_locale = :en
-    Capybara.server = :puma
+  end
+
+  config.around do |example|
+    I18n.with_locale(:en) { example.run }
   end
 end
